+++ src/pages/FindVehicles.jsx
@@ -2,8 +2,12 @@ import React, { useState, useEffect } from 'react'
 import axiosConfig from '../api/axiosConfig';
 import '../styles/VehicleList.css';
 import VehicleList from '../components/VehicleList';
+import { useLocation } from 'react-router-dom';
 
 const FindVehicles = () => {
+    const location = useLocation();
+    const type = location.state
+
     const [vehicles, setVehicles] = useState([]);
     const [loading, setLoading] = useState(true);
 
@@ -19,9 +23,28 @@ const FindVehicles = () => {
         }
     };
 
+    const searchByType = async(vehicleType) => {
+        try {
+          setVehicles([]); 
+          console.log(vehicleType);
+          const response = await axiosConfig.get(`/home/keyword/${vehicleType}`);
+          setVehicles(response.data);
+      } catch (error) {
+          console.error("Error fetching vehicles:", error);
+      } finally {
+          setLoading(false);
+      }
+      };
+
     useEffect(() => {
-        fetchVehicles();
-    }, [])
+        console.log("Received type:", type);
+        if(!type){
+            fetchVehicles();
+        }
+        else{
+            searchByType(type);
+        }
+    }, [type])
     return (
         <div id='vehicle-list-section'>
             {
